Day 5 Part 1 solution
I found it interesting that I had to jump around the type system here.
The instructions are definitely signed integers, and specifying 64-bit
is generally considered good practice. The loop ensures its contents
will only run in the context where coercing an `i64` to `usize` is
safe, but I had to think about this rather than it being enforced. :-(