Day 6 Part A solution
The most curious part of this was having to prefer the first bank if
several had the maximum value. `Iterator::max_by` prefers the last. I
worked around that by reversing the order of the iteration after the
`Iterator::enumerate` call, which works but isn't something I'd care
to maintain. It also isn't tested.

In any real code, and in future even in these small examples, I'd want
to pull out the maximum-bank and rebalancing code into separate
function(s). That would present a better surface for testing.